src/main/java/com/example/dvdrental/api/PaymentsController.java
package com.example.dvdrental.api;

import com.example.dvdrental.repository.PaymentRepository;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/payments")
public class PaymentsController {

    private PaymentRepository paymentRepository;

    public PaymentsController(PaymentRepository paymentRepository) {
        this.paymentRepository = paymentRepository;
    }

    @GetMapping("/total")
    public ResponseEntity getPayments() {
        return ResponseEntity.ok(paymentRepository.totalRentalAmount());
    }
}
